+++ fillit_sources/solve.c
@@ -6,7 +6,7 @@
 /*   By: thakala <thakala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 18:31:19 by mrozhnova         #+#    #+#             */
-/*   Updated: 2022/01/26 15:59:04 by thakala          ###   ########.fr       */
+/*   Updated: 2022/01/26 16:20:12 by thakala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ static char	*ft_strnewset(char chr, uint16_t len)
 
 /* tetriminoes[(uint64_t)depth] too big cast? */
 
-char	*solve(t_tetri *tetriminoes, uint8_t board_size, char depth)
+char	*solve(t_tetri *tetriminoes, uint16_t board_size, char depth)
 {
 	static uint16_t	bitcount;
 	uint16_t		index;
@@ -58,7 +58,7 @@ char	*solve(t_tetri *tetriminoes, uint8_t board_size, char depth)
 	index = 0;
 	while (index < bitcount)
 	{
-		tetrilong = pad_short(tetriminoes[(uint64_t)depth].shape, index, board_size);
+		tetrilong = pad_short(tetriminoes[(uint64_t)depth].shape, index, (uint8_t)board_size);
 		if (tetrilong != (uint64_t)(-1) && bitarrset(bitarray(bitcount, FETCH), index, tetrilong) == 1)
 		{
 			answer = solve(tetriminoes, board_size, depth + 1);
